fix(backend/paddle): address Gemini round 7 review
- convert_to_tensor: pass dtype in recursive list/tuple stacking
- in_top_k: extract values from paddle.topk tuple with [0]
- cdist: dimension-agnostic transpose for y_norm
- quantile: flatten input when axis=None
- dot_product_attention: use paddle.shape for dynamic causal mask

+++ keras/src/backend/paddle/core.py
@@ -162,7 +162,9 @@ def convert_to_tensor(x, dtype=None, sparse=None, ragged=None):
             is_tensor(item) or isinstance(item, Variable)
             for item in tree.flatten(x)
         ):
-            return paddle.stack([convert_to_tensor(x1) for x1 in x])
+            return paddle.stack(
+                [convert_to_tensor(x1, dtype=dtype) for x1 in x]
+            )
     elif not isinstance(x, (bool, int, float)):
         x = np.array(x)
     if isinstance(x, np.ndarray):
